requirements.txt
Flask==3.0.0
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
geopy==2.4.1
pytest==7.4.3
